Move translation components to a folder
