+++ jsx/App/AboutPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { aboutPageJSX } from './LocaleConstants.jsx';
-import { TranslatableText } from './TranslatableText.jsx'
+import { aboutPageJSX } from './locale/LocaleConstants.jsx';
+import { TranslatableText } from './locale/TranslatableText.jsx'
 
 export function AboutPage() {
   return <TranslatableText dictionary={aboutPageJSX} />;
