+++ jsx/App/GlossaryPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { glossaryPageJSX } from './LocaleConstants.jsx';
-import { TranslatableText } from './TranslatableText.jsx'
+import { glossaryPageJSX } from './locale/LocaleConstants.jsx';
+import { TranslatableText } from './locale/TranslatableText.jsx'
 
 export function GlossaryPage() {
   return <TranslatableText dictionary={glossaryPageJSX} />;
