+++ jsx/App/LandingPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { landingPageJSX } from './LocaleConstants.jsx';
-import { TranslatableText } from './TranslatableText.jsx'
+import { landingPageJSX } from './locale/LocaleConstants.jsx';
+import { TranslatableText } from './locale/TranslatableText.jsx'
 
 export function LandingPage() {
   return <TranslatableText dictionary={landingPageJSX} />;
