+++ jsx/App/Stories/Loader.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { loadingPageText } from './../LocaleConstants.jsx';
-import { TranslatableText } from './../TranslatableText.jsx'
+import { loadingPageText } from '../locale/LocaleConstants.jsx';
+import { TranslatableText } from '../locale/TranslatableText.jsx'
 
 export function Loader() {
   return (
