+++ jsx/App/Stories/NotFound.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { notFoundPageText } from './../LocaleConstants.jsx';
-import { TranslatableText } from './../TranslatableText.jsx'
+import { notFoundPageText } from '../locale/LocaleConstants.jsx';
+import { TranslatableText } from '../locale/TranslatableText.jsx'
 
 export function NotFound() {
   return (
