+++ jsx/App/Stories/Story/Sidebar/Minibar/Info/Metadata.jsx
@@ -1,5 +1,5 @@
-import { metadataDescriptionText, metadataAuthorText, metadataGenreText, metadataDateText, metadataGlosserText, metadataSourceText } from './../../../../../LocaleConstants.jsx';
-import { TranslatableText } from './../../../../../TranslatableText.jsx'
+import { metadataDescriptionText, metadataAuthorText, metadataGenreText, metadataDateText, metadataGlosserText, metadataSourceText } from '../../../../../locale/LocaleConstants.jsx';
+import { TranslatableText } from '../../../../../locale/TranslatableText.jsx'
 
 export function Metadata({ metadata }) {
 	// I/P: metadata, in JSON format
