+++ jsx/App/locale/LocaleConstants.jsx
